fix: sc-1489 bug and ui fixes
