+++ shared/components/General/CopyToClipboard/CopyToClipboard.styles.ts
@@ -11,6 +11,18 @@ export const styles = {
     border-radius: 10px;
     display: flex;
     align-items: center;
+    cursor: pointer;
+
+    path {
+      color: ${theme.colorLabel};
+    }
+
+    :hover,
+    :active {
+      path {
+        color: ${theme.colorText};
+      }
+    }
   }`,
   value: (theme: ITheme) => css`
     padding: 5px 10px;
@@ -21,44 +33,4 @@ export const styles = {
     background-color: ${theme.colorInput};
     margin-right: auto;
   }`,
-  button: css`
-    border: none;
-    outline: none;
-    background: none;
-    cursor: pointer;
-    margin-left: auto;
-
-    &::before {
-      content: "Copied";
-      position: absolute;
-      top: -45px;
-      right: 0px;
-      background: #5c81dc;
-      padding: 8px 10px;
-      border-radius: 20px;
-      font-size: 15px;
-      display: none;
-    }
-
-    &::after {
-      content: "";
-      position: absolute;
-      top: -20px;
-      right: 25px;
-      width: 10px;
-      height: 10px;
-      background: #5c81dc;
-      transform: rotate(45deg);
-      display: none;
-    }
-  }`,
-
-  buttonActive: css`
-    ::before {
-      display: block;
-    }
-    ::after {
-      display: block;
-    }
-  `,
 };
