+++ shared/components/General/CopyToClipboard/CopyToClipboard.tsx
@@ -1,17 +1,20 @@
 import { styles } from './CopyToClipboard.styles';
-import { Copy } from '../Copy/Copy';
+import { copyToClipboard } from '@shared/utils/copyToClipboard';
+import CopyIcon from '@public/assets/icons/copy-12.svg';
 
 type CopyToClipboardProps = {
   value: string;
 };
 
 export const CopyToClipboard = ({ value }: CopyToClipboardProps) => {
+  const handleCopy = () => copyToClipboard(value);
+
   return (
     <>
-      <div css={styles.wrapper}>
+      <div css={styles.wrapper} onClick={handleCopy}>
         <span>$</span>
         <p css={styles.value}>{value}</p>
-        <Copy value={value} />
+        <CopyIcon />
       </div>
     </>
   );
