+++ shared/index.ts
@@ -9,3 +9,4 @@ export * from './hooks/useTabs';
 export * from './utils/storybook';
 export * from './utils/escapeHtml';
 export * from './utils/formatters';
+export * from './utils/copyToClipboard';
